[Frontend] merge beam search implementations (#9296)
